Move test_format_time to test_utils
