+++ dask/diagnostics/tests/test_progress.py
@@ -4,7 +4,6 @@ import pytest
 
 from dask.callbacks import Callback
 from dask.diagnostics import ProgressBar
-from dask.diagnostics.progress import format_time
 from dask.local import get_sync
 from dask.threaded import get as get_threaded
 
@@ -56,14 +55,6 @@ def test_clean_exit(get):
         pass
     assert not pbar._running
     assert not pbar._timer.is_alive()
-
-
-def test_format_time():
-    assert format_time(1.4) == " 1.4s"
-    assert format_time(10.4) == "10.4s"
-    assert format_time(100.4) == " 1min 40.4s"
-    assert format_time(1000.4) == "16min 40.4s"
-    assert format_time(10000.4) == " 2hr 46min 40.4s"
 
 
 def test_register(capsys):
