+++ dask/tests/test_utils.py
@@ -25,6 +25,7 @@ from dask.utils import (
     extra_titles,
     factors,
     format_bytes,
+    format_time,
     funcname,
     getargspec,
     has_keyword,
@@ -774,6 +775,16 @@ def test_stringify_collection_keys():
 )
 def test_format_bytes(n, expect):
     assert format_bytes(int(n)) == expect
+
+
+def test_format_time():
+    assert format_time(1.4) == " 1.4s"
+    assert format_time(10.4) == "10.4s"
+    assert format_time(100.4) == " 1min 40.4s"
+    assert format_time(1000.4) == "16min 40.4s"
+    assert format_time(10000.4) == " 2hr 46min 40.4s"
+    assert format_time(1234.567) == "20m 34s"
+    assert format_time(12345.67) == "3h 25m"
 
 
 def test_deprecated():
